+++ src/main/java/online/templab/flippedclass/entity/CourseMemberLimitStrategy.java
@@ -28,13 +28,13 @@ public class CourseMemberLimitStrategy implements Serializable {
      * 队伍中选该课程最少人数
      */
     @Column(name = "`min_member`")
-    private Byte minMember;
+    private Integer minMember;
 
     /**
      * 队伍中选该课程最多人数
      */
     @Column(name = "`max_member`")
-    private Byte maxMember;
+    private Integer maxMember;
 
     private static final long serialVersionUID = 1L;
 }
